requirements.txt
accelerate>=0.31.0
diffusers>=0.31.0
numpy<2.0
torchvision
transformers>=4.41.2
datasets
peft>=0.6.0
ftfy
tensorboard
Jinja2
hpsv2
openai_clip==1.0.1
huggingface_hub>=0.34.0
scikit-image
sentencepiece
